feat!: add icon to external links
BREAKING CHANGE: Renamed changeNavMarkup.js to addIconToExternalLinks.js
