+++ tacc_readthedocs/js/tacc-theme.mjs
@@ -4,4 +4,4 @@ import './tacc-theme/redirectNavLinks.js';
 import './tacc-theme/setTargetForExternalLinks.mjs';
 import './tacc-theme/swapImgSvgWithRawSvg.mjs';
 import './tacc-theme/removeThemeClasses.js';
-import './tacc-theme/changeNavMarkup.js';
+import './tacc-theme/addIconToExternalLinks.js';
